pyi2tun/requirements.txt
pycrypto
enum34>=1.0
pyelliptic==1.5.6
libnacl>=1.4.3
python-pytun
